Recompile report.pdf
TeXpad compiled a version of report.pdf that has extra spacings depending on the original .tex (i.e. it adds extra paragraph spaces in places it shouldn't have) so it has been recompiled with pdflatex.

In the future this file might be added to .gitignore and instead be supplied via Releases.